feat(cpuOnOff): add CPU OnOff module controled by the resource PetriNet
* test(13.1.0): Changes in the toggle_active_cpu to run method on module load directly

This way we prevent to validate if we have to turn on/off the CPU queuing in every transition fire. Also we are doing the fire instantly to improve timing posible issues.

* test(13.1.0): Fix print format for the curthread

* test(13.1.0): Fix wrong parenthesis closure

* feat(idlethread): add new transition TRAN_EXEC_IDLE

With this transition, the idea is to reduce the simulation of getting the idlethread to run. We replace the two transitions that add the thread to the queue and consecutively dequeue it to only one transition that is in charge of setting the idle thread to run in the cpu passed as parameter

* fix(idlethread): use TRAN_EXEC_IDLE by processor

The issue was that we were always trying to run the P0 transition and that was giving transition non sensitized errors (of course :o)

* fix(cpuOnOff): cpu OnOff logic to avoid running uninteded code in sched_choose

In this commit we fix the issue that was causing the suspended processors fire the TRAN_FROM_GLOBAL_CPU when it was inhibited by the PLACE_SUSPENDED. We had this case uncovered and now we are solving it by checking the suspended place tokens and using the idlethread in those cases.

* fix(cpuOnOff): add logic to avoid turning off CPU 0.

Add the  logic to avoid turning off the CPU 0, as we tested it and it breaks the hole system.

* chore(13.1.0): Clean unused code and print functions